生成库文件
