--- SyntacticAnalyzer/SyntacticAnalyzer/SyntacticAnalyzer.qrc
@@ -1,4 +1,0 @@
-<RCC>
-    <qresource prefix="SyntacticAnalyzer">
-    </qresource>
-</RCC>
