--- SyntacticAnalyzer/SyntacticAnalyzer/config.ini
@@ -1,3 +1,0 @@
-[path]
-terminals=E:/CppProject/SyntacticAnalyzer/SyntacticAnalyzer/terminals.txt
-grammars=E:/CppProject/SyntacticAnalyzer/SyntacticAnalyzer/grammar.txt
