--- SyntacticAnalyzer/SyntacticAnalyzer/expressions.txt
@@ -1,5 +1,0 @@
-(i+i*i)*i-i
-(i+i*i*i-i
-(i+i*i)*(i-i)
-(i+(i-i))*i-i
-(i+(i*i))
