--- SyntacticAnalyzer/SyntacticAnalyzer/grammar.txt
@@ -1,3 +1,0 @@
-E->E+T|E-T|T
-T->T*F|T/F|F
-F->(E)|i
