--- SyntacticAnalyzer/SyntacticAnalyzer/main.cpp
@@ -1,9 +1,0 @@
-#include "SyntacticAnalyzer.h"
-#include <QtWidgets/QApplication>
-
-int main(int argc, char *argv[]) {
-    QApplication a(argc, argv);
-    SyntacticAnalyzer w;
-    w.show();
-    return a.exec();
-}
