--- SyntacticAnalyzer/SyntacticAnalyzer/terminals.txt
@@ -1,6 +1,0 @@
-(
-)
-*
-+
--
-/
